Fixed conversion warnings

+++ feature_tests/Fortran/reductions/test_shmem_or_01_int8.f90
@@ -47,9 +47,9 @@ program test_shmem_reduction
   integer,   parameter :: min_npes = 3
   integer,   parameter :: nelems = 10
 
-  integer*4, save      :: src(nelems)
+  integer*8, save      :: src(nelems)
   integer*4, save      :: dest(nelems)
-  integer*4  , save    :: dest_expected(nelems)
+  integer*8  , save    :: dest_expected(nelems)
 
   integer,   save      :: pSync(SHMEM_REDUCE_SYNC_SIZE)
   integer  , save      :: pWrk(SHMEM_REDUCE_MIN_WRKDATA_SIZE)
